Restructure CLI directory
